Created permission seeds
Added the actions and target queries and stored necessary ids in a
viariable. Some permissons were added as a test. Must revise before
merging.

Added permissions to all routes requested

These are just a fraction of all the routes the system will have. This
seed will need to be incremented as more routes are ready to be part of
the permission system.