solving rebase conflicts
Moved logic for a helper file

I still need to iron the logic out to get match all cases properly and
find a way to remove all the redundancy in the code. Also none of the
error messages are right yet.

Several corrections to auth middleware

Corrected related methods in helpers/courseInfo, also changed the logic
in checkCourseStaff (api/auth). Helper methods throw errors when
problems are found trying to get a Course related to the
creation/deletion/edition/retrieval of any method.