Work in Progress with user permissions
Imported the permissionCheck helper file. Permissions are needed for showing or
hiding certain selective processes based on the user and their roles. As
soon as these permissions are implemented the changes in this branch can
be finished.

Admin, common and course specific permissions done

Selective Processes now working for admin, common and all user with
permissions in specific courses. Some changes are necessary to deal with global permissions.

Working on PermissionCheck functions

Ended several tests on permission checking functions, they are now ready
to be used in the selective process api file. One more round to go and
this will be (finally) done.

Using functions from permissionCheck

Cleaned up comments, now the code is using functions from a helper file
where all permission checks are stored. It would be a good idea to
refactor auth api using these functions for the sake of uniformity (and
making auth shorter).