Added migration and model for GraduationTypes
Also added the api endpoint. Still need the other files, and the
migration to add the foreign key to the Course model.

Added api, error and fk migration

Added api and error files for the GraduationType model. Also added the
migration that creates the foreign key to GraduationType inside the
Course model and database.

Added migration and seed file

The migration adds the foreign key to GraduationTypes in the Course
model/db, the seed file adds default graduation types to the db.

Added specific route and corrected typos