Installed packages validator and lodash
Changed error calls-01

The error now forwards a messagem from thrown by the api function. This
was done so that an 'errors' object can be sent, showing the result of
validating all the fields in the request.

Added validations to calls create api

This is a WIP commit. All the validation logic will be in
app/validation/calls.js, and imported into this file. All other
validations will follow this pattern.

Moved call validation to validators folder

Added checks for repeated call numbers, tested check that prevents more
than one call to be open at the same time. There are still some
questions on how to solve errors coming from try-catch blocks inside the
validation file. And the validation function needs to be changed, so
that it can also deal with the update function (using all the checks
from the create function, and adding a check for the primary key/id of
the call being updated).

Reorganized validation order

Database validations, which are more expensive since they take longer,
are at the bottom of the function, and are only run if the related
variables passed preliminary (less expensing, quicker, in-memory)
validations.

Next step is having the validate function work with both CREATE and PUT
requests (with or without the primary key id).

Changed validations to check call update

Call updates have to check if a call exists by id. The next step is
blocking createdAt, updatedAt and other fields the user should not be
able to alter.

Validation working for both create and update

There is still the timestamp blocking to be made, but I'll deal with
that on a feature specific branch.