Added validations to Selecetive Process route
Created a validation file for selective processes and implemented the
checks in the create route for the processes. I still need to implement
it on the update route and check if all rules comform to the requirement
(specially check if I missed any rule, and account for js weird edge
cases).

Finishing new validation for processes

The update route now works with validate, the same way the create route.

modify error number