Created InscriptionEvents
This is just the scaffolding, I still need to finish the files for the
api, errors, routes, change de model if needed and create a validation
before finishing the route.

Finished basic CRUD operations

Create, update, delete and list routes are working correctly know. In
the next stage I'm going to create the validation and customize the
error messages to fit this route better.

Implemented validation

Also added validations to the create route in the api file. Now I just
have to add it to the other routes and check if they work as planned.

Added validation to update route

The next and final step is deciding whether it is necessary to
validate the specific and delete routes (to see if there's an element
with specified id, and if the deletion was completed without any error
poping up).