Added duplicate validation to update route
Also changed how the duplicate validation works, removed mentions
to "item" variable in the duplicate validation, since it only used for sending an id when updating
an existing element (Vacancy). I might remove it in all other functions
where it is not needed before the final commit in this branch.